+++ scripts/items/ItemHolder.gd
@@ -6,6 +6,8 @@ class_name ItemHolder
 @onready var sprite_2d = $Sprite2D
 
 func get_item():
+	if _item.item_type == "collectable":
+		SignalBus.item_collected.emit()
 	return _item
 
 # Called when the node enters the scene tree for the first time.