+++ info/ChangeLog.md
@@ -2,16 +2,26 @@
 
 Where changes are placed
 
-## 0.1.10 
-* Added NPC Placeholder Asset
-* Added NPC IDlE
-* Added NPC Dialogue box
+## 0.2
+
+* created the signal bus
+* Player can now get data from the map
+  * The player's pointer will not fire if disabled
+* Base map script created
+* The Item holder will now emit a signal when it's a collevtable and is collected
+* Created Level 1
 
 ## 0.1.11
+
 * New walk sprites for the player
 * Player head can follow the mouse
 * Slime sprite created
 * Slime sprite changed to new sprites
+
+## 0.1.10 
+* Added NPC Placeholder Asset
+* Added NPC IDlE
+* Added NPC Dialogue box
 
 ## 0.1.9